Pi_server/thorlabs-motor-controller/pyproject.toml
[project]
name = "unitelabs-thorlabs-motor-controller"
version = "0.1.0"
description = "A connector for the thorlabs motor controller built with the UniteLabs CDK."
readme = "README.md"
authors = [{name = "Xiao Huan", email = "xhuan07@gmail.com"},]
maintainers = [
   {name = "Xiao Huan", email ="xhuan07@gmail.com"},
]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
    "Topic :: System :: Hardware",
    "Topic :: System :: Hardware :: Universal Serial Bus (USB)",
]
requires-python=">=3.10,<4.0"
dependencies=[
    "unitelabs-cdk~=0.9.0",
    "unitelabs-bus[usb]~=0.6.0",
    "pylablib-lightweight>=1.4.3",
    "pyft232>=0.12",
    "pyusb>=1.3.1",
    "numba>=0.65.0",
]

[project.urls]
repository = "diao.scs.illinois.edu"
documentation = "diao.scs.illinois.edu/-/README.md"
"Bug Tracker" = "diao.scs.illinois.edu/-/issues"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]
dev = [
    "unitelabs-cdk[dev]",
    "ruff",
    "pre-commit",
    "commitizen",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "--color=yes"

[tool.ruff]
src = ["src", "tests"]
line-length = 120

[tool.ruff.lint]
select = [
    "F",      # pyflakes (ruff default)
    "E",      # pycodestyle: error (ruff default: E4, E7, E9)
    "A",      # flake8-builtins
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "ARG",    # flake8-unused-arguments
    "ASYNC",  # flake8-async
    "D",      # pydocstyle
    "D213",   # gDoc-style tweak: multi-line-summary-second-line
    "D401",   # gDoc-style tweak: non-imperative-mood
    "DOC",    # pydoclint (preview-only)
    "ANN001", # flake8-annotations: missing-type-function-argument
    "ANN201", # flake8-annotations: missing-return-type-undocumented-public-function
    "ANN202", # flake8-annotations: missing-return-type-private-function
    "ANN401", # flake8-annotations: any-type
    "EM",     # flake8-errmsg
    "BLE",    # flake8-blind-except
    "PIE",    # flake8-pie
    "SIM",    # flake8-simplify
    "PERF",   # perflint: performance
    "RET",    # flake8-return
    "RSE",    # flake8-raise
    "RUF",    # ruff-specific-rules
    "ISC",    # flake8-implicit-str-concat
    "RET",    # flake8-return
    "PTH",    # flake8-use-pathlib
    "PLW",    # pylint warning
    "UP",     # py-upgrade
    "ERA",    # no commented-out code
    "FIX",    # no FIXME comments
    "T201",   # no print calls for debugging
    "S101",   # no assert
]
ignore = [
    "D100",    # undocumented-public-module
    "D104",    # undocumented-public-package
    "D105",    # undocumented-magic-method
    "D107",    # undocumented-public-init
    "D202",    # blank-line-after-function
    "D212",    # multi-line-summary-first-line
    "D418",    # overload-with-docstring
    "PERF203", # try-except-in-loop
]

[tool.ruff.lint.per-file-ignores]
# ignore documentation, typing, and assertions in tests
"!src/**.py" = ["D", "ANN", "ARG", "S101"]
# ignore parameter name warnings in feature files
"src/unitelabs/thorlabs_motor_controller/features/**/**.py" =  ["D417"]

[tool.ruff.lint.flake8-annotations]
mypy-init-return = false

[tool.ruff.lint.flake8-comprehensions]
allow-dict-calls-with-keyword-arguments = true

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true

[tool.ruff.lint.pydocstyle]
convention = "google"
ignore-var-parameters = true

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
docstring-code-line-length = 80

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver2"
version_provider = "uv"
major_version_zero = true
bump_message = "chore: bump version $current_version → $new_version"
pre_bump_hooks = [
    "echo '[PRE] CZ PreBumpHook...'",
    "if [ \"$(git rev-parse --abbrev-ref HEAD)\" = \"main\" ]; then uv run --no-sync clog unitelabs.thorlabs_motor_controller -o ./CHANGELOG.md -r; else uv run --no-sync clog unitelabs.thorlabs_motor_controller -o ./CHANGELOG.md; fi",
]
post_bump_hooks = [
    "echo '[POST] CZ PostBumpHook...'",
    "git push origin main --tags --follow-tags --no-verify -o ci.skip",
]

[tool.coverage.report]
exclude_also = [
    "@typing.overload",
    "if typing.TYPE_CHECKING:",
    "\\.\\.\\.",
]

[tool.cruft]
skip = [
    ".venv/",
    ".git/",
    ".ruff_cache/",
    ".pytest_cache/",
    ".env",
    ".uv-cache/",
    "uv.lock",
]

[tool.hatch.build.targets.wheel]
packages = ["src/unitelabs"]

[tool.uv]
constraint-dependencies = ["pyqt5-qt5; sys_platform == 'never'"]
default-groups = "all"
prerelease = "if-necessary-or-explicit"

[[tool.uv.index]]
name = "unitelabs"
url = "https://gitlab.com/api/v4/groups/1009252/-/packages/pypi/simple"
authenticate = "auto"
explicit = true

[tool.uv.sources]
unitelabs-bus = { index = "unitelabs" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

